Refactor: 동시성테스트는 통합테스트로 이동
